[avalonia] improve generated code
